bring back typecheck to file
