+++ front/lib/api/files/client_executable.ts
@@ -1,16 +1,13 @@
 import groupBy from "lodash/groupBy";
 
 import {
+  CREATE_CONTENT_CREATION_FILE_TOOL_NAME,
   EDIT_CONTENT_CREATION_FILE_TOOL_NAME,
   REVERT_CONTENT_CREATION_FILE_TOOL_NAME,
 } from "@app/lib/actions/mcp_internal_actions/servers/content_creation/types";
 import {
   getFileContent,
   getUpdatedContentAndOccurrences,
-  isCreateFileActionOutputType,
-  isCreateFileActionType,
-  isEditFileActionType,
-  isRevertFileActionType,
 } from "@app/lib/api/files/utils";
 import type { Authenticator } from "@app/lib/auth";
 import {
@@ -296,7 +293,68 @@ export async function getClientExecutableFileContent(
   }
 }
 
-export async function isCreateFileAction({
+export function isCreateFileActionType(
+  action: AgentMCPActionModel
+): action is AgentMCPActionModel & {
+  augmentedInputs: {
+    content: string;
+  };
+} {
+  return (
+    action.toolConfiguration.originalName ===
+      CREATE_CONTENT_CREATION_FILE_TOOL_NAME &&
+    typeof action.augmentedInputs.content === "string"
+  );
+}
+
+function isEditFileActionType(
+  action: AgentMCPActionModel
+): action is AgentMCPActionModel & {
+  augmentedInputs: { old_string: string; new_string: string };
+} {
+  return (
+    action.toolConfiguration.originalName ===
+      EDIT_CONTENT_CREATION_FILE_TOOL_NAME &&
+    typeof action.augmentedInputs.old_string === "string" &&
+    typeof action.augmentedInputs.new_string === "string"
+  );
+}
+
+function isRevertFileActionType(
+  action: AgentMCPActionModel
+): action is AgentMCPActionModel & {
+  augmentedInputs: { revertCount?: number };
+} {
+  return (
+    action.toolConfiguration.originalName ===
+      REVERT_CONTENT_CREATION_FILE_TOOL_NAME &&
+    typeof action.augmentedInputs === "object"
+  );
+}
+
+function isCreateFileActionOutputType(
+  output: AgentMCPActionOutputItem
+): output is AgentMCPActionOutputItem & {
+  content: { resource: { fileId: string } };
+} {
+  if (typeof output.content !== "object" || output.content === null) {
+    return false;
+  }
+
+  if (
+    typeof output.content.resource !== "object" ||
+    output.content.resource === null
+  ) {
+    return false;
+  }
+
+  return (
+    "fileId" in output.content.resource &&
+    typeof output.content.resource.fileId === "string"
+  );
+}
+
+async function isCreateFileAction({
   action,
   workspace,
   fileId,
